+++ lib/media_source.dart
@@ -42,7 +42,7 @@
 ///   @override
 ///   Future<FileMediaSource<StickerType>> saveTo(String path) async {
 ///     final file = XFile.fromData(bytes, name: name, path: path);
-///     await PlatformUtils.instance.createDirectoryIfNotExists(path);
+///     await PlatformUtils.instance.ensureParentDirectoryExists(path);
 ///     await file.saveTo(path);
 ///     // Provide your own FileMediaSource implementation for StickerType
 ///     throw UnimplementedError('Provide a FileMediaSource<StickerType>');
@@ -103,7 +103,7 @@
 ///   orElse: () => 'Unknown',
 /// );
 /// ```
-library media_source;
+library;
 
 export 'src/sources/asset_media_source.dart';
 export 'src/sources/file_media_source.dart';